requirements.txt
aniso8601==10.0.0
apispec==6.8.1
astroid==3.3.8
attrs==25.1.0
bcrypt==4.2.1
blinker==1.9.0
Cerberus==1.3.7
cfgv==3.4.0
click==8.1.8
colorama==0.4.6
dill==0.3.9
distlib==0.3.9
filelock==3.17.0
Flask==3.1.0
flask-swagger-ui==4.11.1
greenlet==3.1.1
identify==2.6.7
importlib_resources==6.5.2
iniconfig==2.0.0
isort==6.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
marshmallow==3.26.1
mccabe==0.7.0
mysql-connector==2.2.9
nodeenv==1.9.1
packaging==24.2
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.1.0
PyJWT==2.10.1
pylint==3.3.4
PyMySQL==1.1.1
pytest==8.3.4
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.23.1
six==1.17.0
SQLAlchemy==2.0.38
tomlkit==0.13.2
typing_extensions==4.12.2
virtualenv==20.29.2
webargs==8.6.0
Werkzeug==3.1.3
